+++ CHANGELOG.md
@@ -24,9 +24,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 -->
 
+## Creator: [v0.9.0] - 2026-09-03
+
+### Added
+
+- Added buttons for duplicating elements, quests, answer choices, feature presets, and custom icons
+
 ## Creator: [v0.8.0] - 2026-09-03
 
-### Changed
+### Added
 
 - Added drag-and-drop reordering for elements, quests, answer choices, feature presets, and custom icons